Walk SILBoxType structure by looking at their layout arguments.
In the new SIL box scheme, a box's layout is invariant and parameterized by independent generic arguments, so these arguments are what should be structurally walked when transforming or walking types. Type::transform's interface does not currently provide a reasonable way to resolve Substitutions, unfortunately, but our only immediate need to transform SILBoxTypes is for mapping them in and out of generic contexts, for which we can maintain abstract generic conformances.