docs, Mangling: specification of the new mangling scheme.
These are the main changes:

*) Change the order of the mangling to a post-fix like structure.
This is the biggest change.
It will help to get more common prefixes in the mangled names to optimize the trie in the mach-o object files.
The length of the mangled names will mostly stay the same but the order of 'operands' inside the mangling is more or less reversed.
This change also required to use different 'operator' characters in some cases.

*) Word-substitutions
Similar to the S-substitutions, but finer grained. See section 'Identifiers'.
Reduces the size of mangled names in general.

*) Combined substitutions
A more efficient way to mangle multiple A-substitutions (which were S-substitutions in the old scheme)
Reduces the size of mangled names with lots of substitutions, e.g. specialized functions.

*) Change the '_T' prefix to '_S'
Because it's basically a completely new mangling scheme.