Sema: Use TypeReprs to check accessibility
This eliminates a usage of SubstitutedType.

Unfortunately, we still need a simpler version of the old
check for the case where a 'var' has no explicit type written,
and instead the type is inferred from the initializer
expression.

The check for inferred types of vars no longer looks at
SubstitutedType, so it misses the accessibility of dependent
typealias members. A new test case demonstrates the problem.

Note that dependent typealiases were substituted away as far as
the user is concerned anyway, not showing up in generated
interfaces or diagnostics.

Also the new logic is more accurate in the case where a
TypeRepr is present, which is most of the time.