[Type checker] Don't allow us to resolve bogus types from outside protocol scope.
This is a hack to avoid crashing on invalid code where a protocol is
nested inside some other generic context, and one attempts to use a
generic parameter or associated type from the enclosing generic
context. Any attempts to do so will violate the invariant that
protocols are top-level, which now crashes much more readily.

@slava_pestov is working on a far more principled fix for this area.