[ClangImporter] Split out the accessor parts of importMethodType.
There's really not that much code shared between the accessor and
non-accessor cases in importMethodType, so split them up.

The one change here is that previously a method could /think/ it was
an accessor, but not be treated as one if the property decl has a
different getter or setter. Now such a method is just dropped by the
importer completely. It's intended that this is not a behavioral
change in that no real code should have such an AST.

This is building towards always having a consistent type for property
setters added in handlePropertyRedeclaration, which is
rdar://problem/29422993.