Try again to find a good way to deal with LLVM's abi-breaking.h checks.
Neither of my previous attempts to deal with llvm r288082 have worked
out well. Relying on the build-script is not great and changing the
CMake default does not allow sharing the same version of LLVM to build
both Clang and Swift without also disabling the checks for Clang.
Instead of pursuing more build changes, this just adds weak copies of
the missing symbols, which is another way of disabling the link-time
checks.