Revert "Try again to find a good way to deal with LLVM's abi-breaking.h checks."
This reverts commit 24058898fcdcd00c3c712e299759edeb3872805b.

Nope, still not working. This works fine when I build locally, perhaps
because I did an incremental build. It fails on the bots. The runtime is
broken up into multiple libraries, e.g., libswiftReflection.a, and this
change only put the magic symbols into one file that is not linked
everywhere.