[sil-bug-reducer] Initial version of bug_reducer.
Currently it supports random bug finding and opt pipeline reduction. At some
point in the future it will support:

1. Reducing of optimizer crasher test cases.
2. Reducing optimizer miscompile test cases.

But those are for another time, this is just a first step. Patches welcome = ).
Check out llvm's bugpoint for inspiration.

It also has a basic test of pass pipeline reduction that is triggered via the
validation testing in lit. One thing to note here is that the tool right now
assumes darwin. This can be fixed in the future.