Re-enable coverage_smoke.swift to test bots
We disabled this test because it started failing on the incremental
bots. However, I couldn't reproduce the failure locally with a clean
build. Try to reproduce the failure by enabling the test and running it
through CI.

Context: The specific failure was the llvm-cov detected a malformed
coverage mapping in the 'coverage_smoke' binary. While debugging the
issue, I found that the problem is that a FilenameIndex is invalid
(refer: readIntMax(FilenameIndex, TranslationUnitFilenames.size(), in
llvm's CoverageMappingReader). This means that the swift compiler is
somehow generating an invalid index, although I don't see how this is
possible.