[sil-bug-reducer] Add function reducing functionality.
Now once you have a pass list using:

  ./bug-reducer opt ...

You can take the pass list and give it and the output file to bug-reducer func:

  ./bug-reducer func ...

which will reduce the function list for you given a pass list. This is kept
separate to allow for testing and orthogonality. In a future commit, I am going
to add an option to ./bug-reducer opt that will invoke the function reducer
automagically.