[sil-bug-reducer] Pass in a config object to SILToolInvoker instead of the full args.
This enables us to specify variables like module_name, sdk without having the
SILToolInvoker rely on the argument for the input_file. This is a nice
separation and will let me implement easily a func_bug_reducer tool that can be
enabled after reducing  the number of functions using opt_bug_reducer.py.