[SILOpt] Don't do callee analysis on destructors of imported classes.
If by chance we haven't imported the members of a particular class,
SIL should not fault them in if at all possible.

The test case was a pain to come up with; it involves a class using
a forward-declared protocol that's actually defined in Swift /in a
non-primary file/, where that protocol is never mentioned in the
primary file and the class's members are never accessed otherwise.

rdar://problem/29535170