[semantic-sil] Implement ValueOwnershipKindBuiltinVisitor to compute ValueOwnershipKind for Builtins/llvm intrinsics.
This is just a first pass through. We know that builtins can only be used in the
standard library and trivially so only in non-generic contexts.

This means that anything that I marked trivial that isn't really will be caught
and there are very very few builtins that traffic in ownership, so I can sight
verify them.

rdar://29671437