[dead-function-elim] When deleting functions, use a vector instead of itereating over the module functions twice.
We already know the set of dead functions when we drop the bodies, so instead of
iterating over all of the modules in a function (which is a linked list btw),
just use a vector that we create as we drop the bodies. Then in cases where we
are only eliminating a small amount of functions, we only perform a small amount
of work when deleting function bodies.

This is just a small optimization that I saw as I read through the code to make
sure dead function elimination was an idempotent optimization [which it is].

rdar://29650781